+++ src/brain/apis/indext.test.ts
@@ -3,13 +3,13 @@ import { createSlackAppMention } from '@test/utils/createSlackAppMention';
 import { buildServer } from '@/buildServer';
 import { bolt } from '@api/slack';
 
-import getStats, { OWNERSHIP_FILE_LINK } from './getStats';
+import getStatsMessage, { OWNERSHIP_FILE_LINK } from './getStatsMessage';
 import { apis } from '.';
 
 const STATS_TEXT = 'Some random team stats';
 jest.mock('@api/slack');
 
-jest.mock('./getStats', () =>
+jest.mock('./getStatsMessage', () =>
   jest.fn(() => ({
     message: STATS_TEXT,
     should_show_docs: false,
@@ -36,7 +36,7 @@ describe('slack app', function () {
     );
     expect(response.statusCode).toBe(200);
     expect(bolt.client.chat.postMessage).toHaveBeenCalledTimes(1);
-    expect(getStats).toHaveBeenCalledWith("enterprise");
+    expect(getStatsMessage).toHaveBeenCalledWith("enterprise");
     
     expect(bolt.client.chat.update).toHaveBeenCalledTimes(1);
     expect(bolt.client.chat.update.mock.calls[0][0].blocks[0].text.text).toBe(STATS_TEXT)
